TESTS: removed need for registerAsyncHelper (global)
