+++ tests/acceptance/route-hooks-test.js
@@ -1,7 +1,6 @@
-/*global ajax:true*/
-
 import { test } from 'qunit';
 import moduleForAcceptance from '../../tests/helpers/module-for-acceptance';
+import ajax from '../helpers/ajax';
 
 moduleForAcceptance('Acceptance | route hooks');
